International View through Data Visualization
The aim is to develop this portfolio with material for presenting strategic data analyses and business development procedures relative to a game company's operations in global markets. The information ranges from details of user behavior, revenues realized across different regions, languages used, and the kinds of devices being utilized. In such a context, this analysis would be done in supporting the firm to devise and polish its import-export policy or strategy for better revenues. This helps the firm glean the performance of the games through its market segmentation and evolve useful strategies for increasing its top-line revenues and thereby achieving the scalability of the business worldwide.